separate settings service
