+++ core/src/main/scala/org/broadinstitute/dsde/rawls/webservice/WorkspaceApiServiceV2.scala
@@ -11,7 +11,7 @@ import org.broadinstitute.dsde.rawls.model._
 import org.broadinstitute.dsde.rawls.monitor.migration.MultiregionalBucketMigrationJsonSupport._
 import org.broadinstitute.dsde.rawls.openam.UserInfoDirectives
 import org.broadinstitute.dsde.rawls.webservice.CustomDirectives.addLocationHeader
-import org.broadinstitute.dsde.rawls.workspace.{MultiCloudWorkspaceService, WorkspaceService}
+import org.broadinstitute.dsde.rawls.workspace.{MultiCloudWorkspaceService, WorkspaceService, WorkspaceSettingService}
 import spray.json.DefaultJsonProtocol._
 import spray.json._
 
@@ -23,6 +23,7 @@ trait WorkspaceApiServiceV2 extends UserInfoDirectives {
   val workspaceServiceConstructor: RawlsRequestContext => WorkspaceService
   val multiCloudWorkspaceServiceConstructor: RawlsRequestContext => MultiCloudWorkspaceService
   val bucketMigrationServiceConstructor: RawlsRequestContext => BucketMigrationService
+  val workspaceSettingServiceConstructor: RawlsRequestContext => WorkspaceSettingService
 
   def workspaceRoutesV2(otelContext: Context = Context.root()): server.Route =
     requireUserInfo(Option(otelContext)) { userInfo =>
@@ -91,7 +92,7 @@ trait WorkspaceApiServiceV2 extends UserInfoDirectives {
               pathEndOrSingleSlash {
                 get {
                   complete {
-                    workspaceServiceConstructor(ctx)
+                    workspaceSettingServiceConstructor(ctx)
                       .getWorkspaceSettings(workspaceName)
                       .map(StatusCodes.OK -> _)
                   }
@@ -99,7 +100,7 @@ trait WorkspaceApiServiceV2 extends UserInfoDirectives {
                   put {
                     entity(as[List[WorkspaceSetting]]) { settings =>
                       complete {
-                        workspaceServiceConstructor(ctx)
+                        workspaceSettingServiceConstructor(ctx)
                           .setWorkspaceSettings(workspaceName, settings)
                           .map(StatusCodes.OK -> _)
                       }
