+++ core/src/test/scala/org/broadinstitute/dsde/rawls/webservice/ApiServiceSpec.scala
@@ -54,7 +54,8 @@ import org.broadinstitute.dsde.rawls.workspace.{
   MultiCloudWorkspaceService,
   RawlsWorkspaceAclManager,
   WorkspaceRepository,
-  WorkspaceService
+  WorkspaceService,
+  WorkspaceSettingService
 }
 import org.broadinstitute.dsde.workbench.dataaccess.{NotificationDAO, PubSubNotificationDAO}
 import org.broadinstitute.dsde.workbench.google.mock.{MockGoogleBigQueryDAO, MockGoogleIamDAO, MockGoogleStorageDAO}
@@ -392,6 +393,9 @@ trait ApiServiceSpec
       leonardoDAO,
       workbenchMetricBaseName
     )
+
+    override val workspaceSettingServiceConstructor: RawlsRequestContext => WorkspaceSettingService =
+      new WorkspaceSettingService(_, new WorkspaceRepository(slickDataSource), gcsDAO, samDAO)
 
     override val methodConfigurationServiceConstructor: RawlsRequestContext => MethodConfigurationService =
       MethodConfigurationService.constructor(
