+++ core/src/test/scala/org/broadinstitute/dsde/rawls/workspace/WorkspaceServiceUnitTests.scala
@@ -3,20 +3,12 @@ package org.broadinstitute.dsde.rawls.workspace
 import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.model.headers.OAuth2BearerToken
 import bio.terra.workspace.model.{IamRole, RoleBinding, RoleBindingList}
-import com.google.cloud.storage.BucketInfo.LifecycleRule
-import com.google.cloud.storage.BucketInfo.LifecycleRule.{LifecycleAction, LifecycleCondition}
 import org.broadinstitute.dsde.rawls.billing.{BillingProfileManagerDAO, BillingRepository}
 import org.broadinstitute.dsde.rawls.config._
 import org.broadinstitute.dsde.rawls.dataaccess._
 import org.broadinstitute.dsde.rawls.dataaccess.leonardo.LeonardoService
 import org.broadinstitute.dsde.rawls.dataaccess.workspacemanager.WorkspaceManagerDAO
 import org.broadinstitute.dsde.rawls.fastpass.FastPassServiceImpl
-import org.broadinstitute.dsde.rawls.model.WorkspaceSettingConfig.{
-  GcpBucketLifecycleAction,
-  GcpBucketLifecycleCondition,
-  GcpBucketLifecycleConfig,
-  GcpBucketLifecycleRule
-}
 import org.broadinstitute.dsde.rawls.model.WorkspaceType.WorkspaceType
 import org.broadinstitute.dsde.rawls.model._
 import org.broadinstitute.dsde.rawls.resourcebuffer.ResourceBufferServiceImpl
@@ -38,7 +30,6 @@ import org.mockito.ArgumentMatchers._
 import org.mockito.Mockito._
 import org.scalatest.OptionValues
 import org.scalatest.flatspec.AnyFlatSpec
-import org.scalatest.matchers.must.Matchers.{contain, include}
 import org.scalatest.matchers.should.Matchers.convertToAnyShouldWrapper
 import spray.json.{JsObject, JsString}
 
@@ -799,393 +790,5 @@ class WorkspaceServiceUnitTests extends AnyFlatSpec with OptionValues with Mocki
     )
 
     Await.result(service.updateACL(WorkspaceName("fake_namespace", "fake_name"), aclUpdate, true), Duration.Inf)
-  }
-
-  "getWorkspaceSettings" should "return the workspace settings" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-    val workspaceSettings = List(
-      WorkspaceSetting(WorkspaceSettingTypes.GcpBucketLifecycle,
-                       WorkspaceSettingTypes.GcpBucketLifecycle.defaultConfig()
-      )
-    )
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-    when(workspaceRepository.getWorkspaceSettings(workspaceId)).thenReturn(Future.successful(workspaceSettings))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.read),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val service =
-      workspaceServiceConstructor(samDAO = samDAO, workspaceRepository = workspaceRepository)(defaultRequestContext)
-
-    val returnedSettings = Await.result(service.getWorkspaceSettings(workspaceName), Duration.Inf)
-    returnedSettings shouldEqual workspaceSettings
-    verify(samDAO).userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                                 ArgumentMatchers.eq(workspaceId.toString),
-                                 ArgumentMatchers.eq(SamWorkspaceActions.read),
-                                 any()
-    )
-  }
-
-  it should "handle a workspace with no settings" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-    when(workspaceRepository.getWorkspaceSettings(workspaceId)).thenReturn(Future.successful(List.empty))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.read),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val service =
-      workspaceServiceConstructor(samDAO = samDAO, workspaceRepository = workspaceRepository)(defaultRequestContext)
-
-    val returnedSettings = Await.result(service.getWorkspaceSettings(workspaceName), Duration.Inf)
-    returnedSettings shouldEqual List.empty
-    verify(samDAO).userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                                 ArgumentMatchers.eq(workspaceId.toString),
-                                 ArgumentMatchers.eq(SamWorkspaceActions.read),
-                                 any()
-    )
-  }
-
-  it should "return an error if the user does not have read access to the workspace" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.read),
-                           any()
-      )
-    ).thenReturn(Future.successful(false))
-    val service =
-      workspaceServiceConstructor(samDAO = samDAO, workspaceRepository = workspaceRepository)(defaultRequestContext)
-
-    assertThrows[NoSuchWorkspaceException] {
-      Await.result(service.getWorkspaceSettings(workspaceName), Duration.Inf)
-    }
-  }
-
-  "setWorkspaceSettings" should "set the workspace settings if there aren't any set" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-    val workspaceSetting = WorkspaceSetting(WorkspaceSettingTypes.GcpBucketLifecycle,
-                                            WorkspaceSettingTypes.GcpBucketLifecycle.defaultConfig()
-    )
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-    when(workspaceRepository.getWorkspaceSettings(workspaceId)).thenReturn(Future.successful(List.empty))
-    when(
-      workspaceRepository.createWorkspaceSettingsRecords(workspaceId,
-                                                         List(workspaceSetting),
-                                                         defaultRequestContext.userInfo.userSubjectId
-      )
-    )
-      .thenReturn(Future.successful(List(workspaceSetting)))
-    when(workspaceRepository.markWorkspaceSettingApplied(workspaceId, workspaceSetting.settingType))
-      .thenReturn(Future.successful(1))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.own),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val gcsDAO = mock[GoogleServicesDAO]
-    when(gcsDAO.setBucketLifecycle(workspace.bucketName, List())).thenReturn(Future.successful())
-
-    val service = workspaceServiceConstructor(samDAO = samDAO,
-                                              workspaceRepository = workspaceRepository,
-                                              gcsDAO = gcsDAO
-    )(defaultRequestContext)
-
-    val res = Await.result(service.setWorkspaceSettings(workspaceName, List(workspaceSetting)), Duration.Inf)
-    res.successes should contain theSameElementsAs List(workspaceSetting)
-    res.failures shouldEqual Map.empty
-  }
-
-  it should "overwrite existing settings" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-    val existingSetting = WorkspaceSetting(
-      WorkspaceSettingTypes.GcpBucketLifecycle,
-      GcpBucketLifecycleConfig(
-        List(
-          GcpBucketLifecycleRule(GcpBucketLifecycleAction("Delete"),
-                                 GcpBucketLifecycleCondition(Some(Set("prefixToMatch")), Some(30))
-          )
-        )
-      )
-    )
-    val newSetting = WorkspaceSetting(
-      WorkspaceSettingTypes.GcpBucketLifecycle,
-      GcpBucketLifecycleConfig(
-        List(
-          GcpBucketLifecycleRule(GcpBucketLifecycleAction("Delete"),
-                                 GcpBucketLifecycleCondition(Some(Set("muchBetterPrefix")), Some(31))
-          )
-        )
-      )
-    )
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-    when(workspaceRepository.getWorkspaceSettings(workspaceId)).thenReturn(Future.successful(List(existingSetting)))
-    when(
-      workspaceRepository.createWorkspaceSettingsRecords(workspaceId,
-                                                         List(newSetting),
-                                                         defaultRequestContext.userInfo.userSubjectId
-      )
-    )
-      .thenReturn(Future.successful(List(newSetting)))
-    when(workspaceRepository.markWorkspaceSettingApplied(workspaceId, newSetting.settingType))
-      .thenReturn(Future.successful(1))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.own),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val gcsDAO = mock[GoogleServicesDAO]
-    val newSettingGoogleRule = new LifecycleRule(
-      LifecycleAction.newDeleteAction(),
-      LifecycleCondition.newBuilder().setMatchesPrefix(List("muchBetterPrefix").asJava).setAge(31).build()
-    )
-    when(gcsDAO.setBucketLifecycle(workspace.bucketName, List(newSettingGoogleRule))).thenReturn(Future.successful())
-
-    val service = workspaceServiceConstructor(samDAO = samDAO,
-                                              workspaceRepository = workspaceRepository,
-                                              gcsDAO = gcsDAO
-    )(defaultRequestContext)
-
-    val res = Await.result(service.setWorkspaceSettings(workspaceName, List(newSetting)), Duration.Inf)
-    res.successes should contain theSameElementsAs List(newSetting)
-    res.failures shouldEqual Map.empty
-  }
-
-  it should "remove existing settings if no settings are specified" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-    val existingSetting = WorkspaceSetting(
-      WorkspaceSettingTypes.GcpBucketLifecycle,
-      GcpBucketLifecycleConfig(
-        List(
-          GcpBucketLifecycleRule(GcpBucketLifecycleAction("Delete"),
-                                 GcpBucketLifecycleCondition(Some(Set("prefixToMatch")), Some(30))
-          )
-        )
-      )
-    )
-    val defaultSetting = WorkspaceSetting(WorkspaceSettingTypes.GcpBucketLifecycle,
-                                          WorkspaceSettingTypes.GcpBucketLifecycle.defaultConfig()
-    )
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-    when(workspaceRepository.getWorkspaceSettings(workspaceId)).thenReturn(Future.successful(List(existingSetting)))
-    when(
-      workspaceRepository.createWorkspaceSettingsRecords(workspaceId,
-                                                         List.empty,
-                                                         defaultRequestContext.userInfo.userSubjectId
-      )
-    )
-      .thenReturn(Future.successful(List.empty))
-    when(workspaceRepository.markWorkspaceSettingApplied(workspaceId, defaultSetting.settingType))
-      .thenReturn(Future.successful(1))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.own),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val gcsDAO = mock[GoogleServicesDAO]
-    when(gcsDAO.setBucketLifecycle(workspace.bucketName, List())).thenReturn(Future.successful())
-
-    val service = workspaceServiceConstructor(samDAO = samDAO,
-                                              workspaceRepository = workspaceRepository,
-                                              gcsDAO = gcsDAO
-    )(defaultRequestContext)
-
-    val res = Await.result(service.setWorkspaceSettings(workspaceName, List.empty), Duration.Inf)
-    res.successes should contain theSameElementsAs List(defaultSetting)
-    res.failures shouldEqual Map.empty
-  }
-
-  it should "report errors while applying settings and remove pending settings" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-    val existingSetting = WorkspaceSetting(
-      WorkspaceSettingTypes.GcpBucketLifecycle,
-      GcpBucketLifecycleConfig(
-        List(
-          GcpBucketLifecycleRule(GcpBucketLifecycleAction("Delete"),
-                                 GcpBucketLifecycleCondition(Some(Set("prefixToMatch")), Some(30))
-          )
-        )
-      )
-    )
-    val newSetting = WorkspaceSetting(
-      WorkspaceSettingTypes.GcpBucketLifecycle,
-      GcpBucketLifecycleConfig(
-        List(
-          GcpBucketLifecycleRule(GcpBucketLifecycleAction("Delete"),
-                                 GcpBucketLifecycleCondition(Some(Set("muchBetterPrefix")), Some(31))
-          )
-        )
-      )
-    )
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-    when(workspaceRepository.getWorkspaceSettings(workspaceId)).thenReturn(Future.successful(List(existingSetting)))
-    when(
-      workspaceRepository.createWorkspaceSettingsRecords(workspaceId,
-                                                         List(newSetting),
-                                                         defaultRequestContext.userInfo.userSubjectId
-      )
-    )
-      .thenReturn(Future.successful(List(newSetting)))
-    when(workspaceRepository.removePendingSetting(workspaceId, newSetting.settingType)).thenReturn(Future.successful(1))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.own),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val gcsDAO = mock[GoogleServicesDAO]
-    val newSettingGoogleRule = new LifecycleRule(
-      LifecycleAction.newDeleteAction(),
-      LifecycleCondition.newBuilder().setMatchesPrefix(List("muchBetterPrefix").asJava).setAge(31).build()
-    )
-    when(gcsDAO.setBucketLifecycle(workspace.bucketName, List(newSettingGoogleRule)))
-      .thenReturn(Future.failed(new Exception("failed to apply settings")))
-
-    val service = workspaceServiceConstructor(samDAO = samDAO,
-                                              workspaceRepository = workspaceRepository,
-                                              gcsDAO = gcsDAO
-    )(defaultRequestContext)
-
-    val res = Await.result(service.setWorkspaceSettings(workspaceName, List(newSetting)), Duration.Inf)
-    res.successes shouldEqual List.empty
-    res.failures(WorkspaceSettingTypes.GcpBucketLifecycle) shouldEqual ErrorReport(StatusCodes.InternalServerError,
-                                                                                   "failed to apply settings"
-    )
-    verify(workspaceRepository).removePendingSetting(workspaceId, newSetting.settingType)
-  }
-
-  it should "be limited to owners" in {
-    val workspaceId = workspace.workspaceIdAsUUID
-    val workspaceName = workspace.toWorkspaceName
-    val newSetting = WorkspaceSetting(WorkspaceSettingTypes.GcpBucketLifecycle,
-                                      WorkspaceSettingTypes.GcpBucketLifecycle.defaultConfig()
-    )
-
-    val workspaceRepository = mock[WorkspaceRepository]
-    when(workspaceRepository.getWorkspace(workspaceName, None)).thenReturn(Future.successful(Option(workspace)))
-
-    val samDAO = mock[SamDAO]
-    when(samDAO.getUserStatus(any()))
-      .thenReturn(Future.successful(Option(SamUserStatusResponse("fake_user_id", "user@example.com", true))))
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.own),
-                           any()
-      )
-    ).thenReturn(Future.successful(false))
-    // Rawls confirms a user has at least read access to the workspace after a failed authz check
-    // to determine if it should throw a 403 or 404
-    when(
-      samDAO.userHasAction(ArgumentMatchers.eq(SamResourceTypeNames.workspace),
-                           ArgumentMatchers.eq(workspaceId.toString),
-                           ArgumentMatchers.eq(SamWorkspaceActions.read),
-                           any()
-      )
-    ).thenReturn(Future.successful(true))
-
-    val service =
-      workspaceServiceConstructor(samDAO = samDAO, workspaceRepository = workspaceRepository)(defaultRequestContext)
-
-    val exception = intercept[RawlsExceptionWithErrorReport] {
-      Await.result(service.setWorkspaceSettings(workspaceName, List(newSetting)), Duration.Inf)
-    }
-    exception.errorReport.statusCode shouldBe Some(StatusCodes.Forbidden)
-  }
-
-  it should "validate requested settings" in {
-    val workspaceName = workspace.toWorkspaceName
-    val newSetting = WorkspaceSetting(
-      WorkspaceSettingTypes.GcpBucketLifecycle,
-      GcpBucketLifecycleConfig(
-        List(
-          GcpBucketLifecycleRule(GcpBucketLifecycleAction("SetStorageClass"),
-                                 GcpBucketLifecycleCondition(Some(Set("prefixToMatch")), Some(-1))
-          )
-        )
-      )
-    )
-
-    val service = workspaceServiceConstructor()(defaultRequestContext)
-
-    val exception = intercept[RawlsExceptionWithErrorReport] {
-      Await.result(service.setWorkspaceSettings(workspaceName, List(newSetting)), Duration.Inf)
-    }
-    exception.errorReport.statusCode shouldBe Some(StatusCodes.BadRequest)
-    exception.errorReport.message should include("Invalid settings requested.")
-    exception.errorReport.causes should contain theSameElementsAs List(
-      ErrorReport("Invalid GcpBucketLifecycle configuration: age must be a non-negative integer."),
-      ErrorReport("Invalid GcpBucketLifecycle configuration: unsupported lifecycle action SetStorageClass.")
-    )
   }
 }
